emission-api/requirements.txt
Django==5.1.7
djangorestframework==3.15.2
django-cors-headers==4.5.0
pytz==2021.1
flake8==7.0.0
